+++ plugins/actionlog/joomla/services/provider.php
@@ -15,7 +15,7 @@ use Joomla\Database\DatabaseInterface;
 use Joomla\DI\Container;
 use Joomla\DI\ServiceProviderInterface;
 use Joomla\Event\DispatcherInterface;
-use Joomla\Plugin\Actionlog\Joomla\Joomla;
+use Joomla\Plugin\Actionlog\Joomla\Joomla\Extension;
 
 return new class implements ServiceProviderInterface
 {
