+++ plugins/actionlog/joomla/src/Extension/Joomla.php
@@ -7,7 +7,7 @@
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
-namespace Joomla\Plugin\Actionlog\Joomla;
+namespace Joomla\Plugin\Actionlog\Joomla\Extension;
 
 defined('_JEXEC') or die;
